+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Changes:
 
+-   Alert component add
 -   AccountsManager widget styles fix
 
 ## 0.0.35 August 19, 2022